add query without polling
